Added the LocationSelection class to enable searching for locations by name with a GeoCoder(Still a WIP).
Re-mapped the activity_main.xml file so that I can act on it with the fragment transactions.  This allows me to modify the main content view so that I can assign editing views.

Created all the bits and parts to enable the ability to add/remove/update fences on the screen.  Also added the rendering code to the screen so that the circles and markers are drawn.

Added more methods to the CallBack interfaces so that I can tie events happening in the drawer to actions and Fences appearing on the sreen on the Map.